backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
biopython==1.83
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.1
pydantic==2.5.3
python-multipart==0.0.6
reportlab==4.0.8
aiofiles==23.2.1
